Move statistics back to client-side
