+++ app/index.html
@@ -97,6 +97,7 @@
 <script src="team/team.js"></script>
 <script src="addTeam/addTeam.js"></script>
 <script src="services/teamService.js"></script>
+<script src="services/statsService.js"></script>
 <script src="services/userService.js"></script>
 <script src="services/loggedinCheck.js"></script>
 <script src="services/permissions.js"></script>
