--- functions/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "functions",
-  "description": "Cloud Functions for Firebase",
-  "dependencies": {
-    "firebase-admin": "~4.2.1",
-    "firebase-functions": "^0.5.7"
-  },
-  "private": true
-}
